avoid TypeErrors in Page::get_by_template()
If a page with `example-template.php` does not exist, `Page::get_by_template("example-template.php")` returns null. This has resulted in `TypeError`s in production because the return type was declared as `Page`.

Can't use union types here, because PHP (thanks, Obama!) so instead we're just removing the return typehint.