Update PHP and JavaScript dependencies (#156)
* ran composer update for packages

* Update composer to prefer stable versions over branches, and update groot dev dependency to install using the master branch to work around wonky release numbers

* deps: remove duplicate json key

* git: update .gitignore

* deps: replace paultehwalton/acf-stubs

Package is no longer maintained, switched to php-stubs/acf-pro-stubs.

* deps: update phpunit/phpunit to v9

* deps: remove acobster/wp-cli-yaml-fixtures

This package is horrendously out of date since Coby left SiteCrafting,
and it isn't being maintained anymore.

Felt like removing the package was the best path forward instead of
forking it and having to update it.

It has a conflict with behat/behat, as they both require `symfony/yaml`,
 but this acobster/wp-cli-yaml-fixtures wants `symfony/yaml:3.4` and the
  newer versions of `behat/behat` want `symfony/yaml:7.x`.

* deps: update `behat/behat`

Fix an "unmaintained packaged" warning when running `lando composer
audit`.

* fix: errors from `lando sniff`

* deps: update szepeviktor/phpstan-wordpress to v2

* fix: PHPStan errors

* deps: set composer.json and .lando.yml to PHP 8.1

* deps: update johnpbloch/wordpress-core

Set it to be ^6.5.5 so it can pull the latest minor version on composer
updates.

* deps: update wp-coding-standards/wpcs to v3

Remove squizlabs/php_codesniffer because wp-coding-standards/wpcs
requires it, so it makes little sense for us to also require it
when we're not using it directly.

* deps: update johnpbloch/wordpress-core-installer

It looks like johnpbloch/wordpress-core-installer isn't really being
maintained anymore, but it was safe to move off the `dev-master` branch
and onto the `2.0.0` branch instead.

The only reason a v2 was released was because the minimum PHP version
got bumped up to PHP 5.6, which doesn't affect us.

* deps: remove gitbook-cli

This package is no longer maintained and no longer works on Node
versions 10 through 22.

After talking to Scott about it, we decided it was better to just strip
it out for now, as it's probably seldom being used.

* deps: remove victorjonsson/markdowndocs

We were only using this to dynamically generate the API Reference
section of the documentation website.
However, this package, and the others like it, are seldom updated, so it
 was more hassle than it was worth to try and keep it.

* deps: replace gitbook-cli with VitePress

Now the documentation site is blazingly fast 🦀

* docs: minor changes

* feat: add Rector

* misc: ran Rector

* fix: CoPilot feedback where it made sense.

* fix: revert build -> run in .lando.yml

* fix: flesh out "features" on the home page

* fix: sniff fixes

There are some lingering sniff "issues" that I don't consider issues.
But the bulk of the problems have been solved.

---------

Co-authored-by: Reena Hensley <rhensley@sitecrafting.com>
Co-authored-by: Scott Dunham <sdunham@sitecrafting.com>